+++ NEWS.md
@@ -10,6 +10,10 @@ increased.
 
 * Apply MIT license
 
+* Fix bug in `getResultsIntersection()` when a modelID has no features table.
+The "Set_Membership" column was returned in the first column instead of after
+the featureID column (reported by Joe Dalen and Paul Nordlund)
+
 # 0.22.5
 
 * The release tarball includes version 0.3.5.3 of the web app